+++ src/Commands/CurrencyRateCommand.php
@@ -2,7 +2,6 @@
 
 namespace FlexMindSoftware\CurrencyRate\Commands;
 
-use Carbon\Carbon;
 use Illuminate\Console\Command;
 
 class CurrencyRateCommand extends Command
@@ -23,12 +22,8 @@ class CurrencyRateCommand extends Command
             $driver = config('currency-rate.driver');
         }
 
-        if (! in_array($driver, array_keys(config('currency-rate.drivers')))) {
-            $this->error('Driver "'.$driver.'" not exists!');
+        $date = new DateTime("@$timestamp");
 
-            return;
-        }
-
-        \CurrencyRate::driver($driver)->downloadRates(Carbon::createFromTimestamp($timestamp));
+        \CurrencyRate::driver($driver)->downloadRates($date);
     }
 }