+++ src/Drivers/BankOfPolandDriver.php
@@ -8,12 +8,13 @@ use FlexMindSoftware\CurrencyRate\Contracts\CurrencyInterface;
 use FlexMindSoftware\CurrencyRate\Models\Currency;
 use FlexMindSoftware\CurrencyRate\Models\CurrencyRate;
 use FlexMindSoftware\CurrencyRate\Models\RateTrait;
-use Log;
 use SimpleXMLElement;
 
 class BankOfPolandDriver extends BaseDriver implements CurrencyInterface
 {
     use RateTrait;
+
+    public const URI = 'https://www.nbp.pl/kursy/xml/';
 
     /**
      * @var string
@@ -48,15 +49,13 @@ class BankOfPolandDriver extends BaseDriver implements CurrencyInterface
 
         $date = date('ymd', $timestamp);
 
-        $config = $this->config['drivers'][$this->driverAlias];
-
-        $listOfCurses = file_get_contents($config['url'] . 'dir.txt');
+        $listOfCurses = file_get_contents(static::URI . 'dir.txt');
 
         if (preg_match_all('/(a)([0-9]{3})z' . $date . '/', $listOfCurses, $matches)) {
             if (! blank($matches[0])) {
                 foreach ($matches[0] as $match) {
                     $nbpNo = $match;
-                    $xml = file_get_contents($config['url'] . $nbpNo . '.xml');
+                    $xml = file_get_contents(static::URI . $nbpNo . '.xml');
                     if (! empty($xml)) {
                         $currencies = new SimpleXMLElement($xml);
                         $currencies = json_decode(json_encode($currencies), true);
@@ -88,13 +87,9 @@ class BankOfPolandDriver extends BaseDriver implements CurrencyInterface
                         }
 
                         CurrencyRate::upsert($toSave, ['no', 'driver', 'code', 'date'], ['rate', 'multiplier']);
-                    } else {
-                        Log::error('No XML: ' . $nbpNo . '.xml');
                     }
                 }
             }
-        } else {
-            Log::error('No NBP currency for: ' . date('Y-m-d', $timestamp));
         }
     }
 }