+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to `flexmind-software/currency-rate` will be documented in this file.
 
-## 0.0.1 - 2021-09-14
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.0.6] - 2021-09-15
+
+### Added
+- add support for Bank of Bulgaria, Bank of Canada, Bank of Czech Republic and European Central Bank
+
+### Changed
+- rebuild driver manager for dynamic call drivers
+
+### Removed
+- nestbot/carbon library
+
+## [0.0.1] - 2021-09-14
 
 - initial release